finished with the api project, Was really fulfilling

+++ task-management/controllers/tasks.js
@@ -2,10 +2,15 @@ import Task from "../models/tasks.js";
 import errorHandler from "../utils/error.js";
 
 const addTask = async (req, res, next) => {
+    const { userId } = req.params;
     const task = req.body;
-    const newTask = new Task(task);
+    const newTask = new Task({ ...task, taskOwner: userId });
 
     try {
+        if (!req.user.admin) {
+            return next(errorHandler(401, "You are not authorized to add tasks, only admins can"))
+        }
+
         await newTask.save();
         res.status(201).json(newTask);
     } catch (err) {
@@ -43,11 +48,25 @@ const updateTask = async (req, res, next) => {
     const { id: taskId } = req.params;
 
     try {
-        const task = await Task.findOneAndUpdate({ _id: taskId }, req.body, {
-            new: true
-        });
+        const task = await Task.findById(taskId);
 
-        res.status(200).json(task);
+        if (!task) {
+            next(errorHandler(404, `There is no task with ID: ${taskId}`));
+        }
+
+        console.log(req.user.id, req.user.taskOwner)
+
+        if (req.user.id == task.taskOwner || req.user.admin) {
+            const task = await Task.findOneAndUpdate({ _id: taskId }, req.body, {
+                new: true
+            });
+
+            return res.status(200).json(task);
+        } else if (req.user.id != task.taskOwner || !req.user.admin) {
+            return next(errorHandler(401, "You are not authorized to update another person's task"));
+        }
+
+
     } catch (err) {
         next(err);
     }
@@ -63,8 +82,14 @@ const deleteTask = async (req, res, next) => {
             return next(errorHandler(404, `The task with ID ${taskId} that you are trying to delete does not exist!`));
         }
 
-        await Task.findByIdAndDelete(taskId);
-        res.status(200).json("task deleted successfully");
+        if (req.user.id == task.taskOwner || req.user.admin) {
+            await Task.findByIdAndDelete(taskId);
+            return res.status(200).json("task deleted successfully");
+        } else if (req.user.id != task.taskOwner || !req.user.admin) {
+            return next(errorHandler(401, "You are not authorized to delete another persons's task"));
+        }
+
+
     } catch (err) {
         next(err)
     }
